+++ crates/wasi-common/src/old/snapshot_0/sys/unix/emscripten/filetime.rs
