+++ crates/wasi-common/src/old/snapshot_0/sys/unix/emscripten/mod.rs
@@ -1,6 +1,6 @@
 pub(crate) mod filetime;
 pub(crate) mod host_impl;
+#[path = "../linux/hostcalls_impl.rs"]
 pub(crate) mod hostcalls_impl;
+#[path = "../linux/oshandle.rs"]
 pub(crate) mod oshandle;
-#[path = "../linux/utimesat.rs"]
-pub(crate) mod utimesat;
