+++ crates/wasi-common/src/old/snapshot_0/sys/unix/linux/oshandle.rs
