+++ crates/wasi-common/src/sys/unix/emscripten/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod filetime;
+#[path = "../linux/host_impl.rs"]
 pub(crate) mod host_impl;
+#[path = "../linux/hostcalls_impl.rs"]
 pub(crate) mod hostcalls_impl;
+#[path = "../linux/oshandle.rs"]
 pub(crate) mod oshandle;
-#[path = "../linux/utimesat.rs"]
-pub(crate) mod utimesat;
