+++ crates/wasi-common/src/sys/unix/linux/hostcalls_impl.rs
