+++ crates/wasi-common/src/sys/unix/linux/mod.rs
@@ -2,5 +2,4 @@ pub(crate) mod filetime;
 pub(crate) mod host_impl;
 pub(crate) mod hostcalls_impl;
 pub(crate) mod oshandle;
-#[path = "../linux/utimesat.rs"]
 pub(crate) mod utimesat;
