+++ crates/wasi-common/src/sys/unix/linux/oshandle.rs
