--- crates/wasi-common/src/sys/unix/linux_like/emscripten/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod filetime;
