--- crates/wasi-common/src/sys/unix/linux_like/linux/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod filetime;
