+++ crates/wasi-common/yanix/src/sys/emscripten/mod.rs
@@ -1,3 +1,10 @@
+#[path = "../linux/dir.rs"]
+pub(crate) mod dir;
+#[path = "../linux/fadvise.rs"]
+pub(crate) mod fadvise;
+#[path = "../linux/file.rs"]
+pub(crate) mod file;
+
 use crate::{dir::SeekLoc, Result};
 use std::convert::TryInto;
