+++ crates/wasi-common/yanix/src/sys/linux/mod.rs
@@ -1,3 +1,7 @@
+pub(crate) mod dir;
+pub(crate) mod fadvise;
+pub(crate) mod file;
+
 use crate::{dir::SeekLoc, Result};
 
 impl SeekLoc {
